[CARBONDATA-3309] MV datamap supports Spark 2.1
